servers/nano-banana-mcp/pyproject.toml
[project]
name = "nano-banana-mcp"
version = "0.1.0"
description = "MCP server for image generation using Google Gemini"
license = { text = "MIT" }
authors = [{ name = "TomorrowToday" }]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.0,<4.0",
    "google-genai>=1.0,<2.0",
    "python-dotenv>=1.0,<2.0",
]

[project.scripts]
nano-banana-mcp = "nano_banana_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
